refactor(*): remove explicit AppContent component
