--- src/components/app-content/index.ts
@@ -1,1 +1,0 @@
-export { AppContent } from "./app-content";
