--- src/components/app-content/style.module.css
@@ -1,5 +1,0 @@
-.main {
-    display: flex;
-    justify-content: center;
-    margin: auto;
-}
