+++ src/components/app/app.module.css
@@ -1,3 +1,9 @@
+.main {
+    display: flex;
+    justify-content: center;
+    margin: auto;
+}
+
 /* #root {
   max-width: 1280px;
   margin: 0 auto;
